fix: move borders module to separate file and add tests
